Introduce exposure history context (#1016)
Why:
A previous commit added the ability to see exposure history data in the
exposure history screen for the bt build. Given that the gps and bt
versions of the app will eventually have the same notion of the exposure
history, we would like to also include this behavior in the gps build.

This commit:
Introduces an exposure history context which takes the logic for
exposing the exposure history from the ExpsoureNotificationsContext such
that we can reuse it in the gps side. We removed the gps specific code
from the ExposureHistory screen such that we no longer needed to do any
data manipulation in the user facing screen components, which allows us
to reuse the screen for both builds and in general provides for a better
separation of view and model logic.
