+++ app/helpers/index.ts
@@ -1,0 +1,3 @@
+import * as DateTimeUtils from './dateTimeUtils';
+
+export { DateTimeUtils };
