+++ app/views/ExposureHistory/ExposureDatumDetail.tsx
@@ -3,13 +3,14 @@ import { TouchableOpacity, StyleSheet, View } from 'react-native';
 import { useNavigation } from '@react-navigation/native';
 import dayjs from 'dayjs';
 
-import { Typography } from '../../components/Typography';
 import {
   ExposureDatum,
   Possible,
   Unknown,
   NoKnown,
-} from '../../ExposureNotificationContext';
+} from '../../ExposureHistoryContext';
+import { Typography } from '../../components/Typography';
+import { TimeHelpers } from '../utils';
 import { Screens } from '../../navigation';
 
 import {
@@ -47,11 +48,14 @@ interface PossibleExposureDetailProps {
 const PossibleExposureDetail = ({
   datum: { date, possibleExposureTimeInMin, currentDailyReports },
 }: PossibleExposureDetailProps) => {
+  const exposureDurationText = TimeHelpers.durationMsToString(
+    possibleExposureTimeInMin * 60 * 1000,
+  );
   const navigation = useNavigation();
   const exposureDate = dayjs(date).format('dddd, MMM DD');
-  const exposureTime = `Possible Exposure Time: ${possibleExposureTimeInMin}`;
+  const exposureTime = `Possible Exposure Time: ${exposureDurationText}`;
   const dailyReports = `Current daily reports: ${currentDailyReports}`;
-  const explainationContent = `For ${possibleExposureTimeInMin} consecutive minutes, your phone was within 10 feet of someone who later received a confirmed positive COVID-19 diagnosis.`;
+  const explainationContent = `For ${exposureDurationText}, your phone was within 10 feet of someone who later received a confirmed positive COVID-19 diagnosis.`;
 
   const handleOnPressNextSteps = () => {
     navigation.navigate(Screens.NextSteps);
@@ -69,13 +73,15 @@ const PossibleExposureDetail = ({
           <Typography style={styles.content}>{explainationContent}</Typography>
         </View>
       </View>
-      <TouchableOpacity
-        style={styles.nextStepsButton}
-        onPress={handleOnPressNextSteps}>
-        <Typography style={styles.nextStepsButtonText}>
-          {nextStepsButtonText}
-        </Typography>
-      </TouchableOpacity>
+      <View style={styles.ctaContainer}>
+        <TouchableOpacity
+          style={styles.nextStepsButton}
+          onPress={handleOnPressNextSteps}>
+          <Typography style={styles.nextStepsButtonText}>
+            {nextStepsButtonText}
+          </Typography>
+        </TouchableOpacity>
+      </View>
     </>
   );
 };
@@ -143,6 +149,9 @@ const styles = StyleSheet.create({
   },
   content: {
     ...TypographyStyles.secondaryContent,
+  },
+  ctaContainer: {
+    marginBottom: Spacing.medium,
   },
   nextStepsButton: {
     ...Buttons.largeBlueOutline,
