+++ app/views/Settings/ENDebugMenu.tsx
@@ -22,7 +22,7 @@ import {
   simulateExposureDetectionError,
   getExposureConfiguration,
 } from '../../exposureNotificationsNativeModule';
-import ExposureNotificationContext from '../../ExposureNotificationContext';
+import ExposureNotificationContext from '../../ExposureHistoryContext';
 import { NavigationProp, Screens } from '../../navigation';
 
 import { Colors, Spacing } from '../../styles';
