+++ app/views/utils/index.ts
@@ -1,0 +1,3 @@
+import * as TimeHelpers from './timeHelpers';
+
+export { TimeHelpers };
